requirements.txt
scrapy==0.24
facebook-sdk